Fix weird races when pasting
The symptom was partial selection when pasting a set of files.

Sometimes when pasting we were allocating two files at the same time
because the a dir reload was racing with a monitor callback, and this
broke the assumption that a ThunarFile instance points to a unique path
(file pointers are used as keys in lookup/comparison).

Other times the initial dir list from the Ctrl-V was so early it ran
before the files appeared, but completed after the monitor callbacks,
causing the files to be removed from the cache before the pasted items
were selected.